--- keyboards/barbellboards/rollow/config.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#define MASTER_LEFT
