--- keyboards/barbellboards/rollow/keymaps/t4corun/config.h
@@ -1,1 +1,0 @@
-#pragma once
