--- keyboards/barbellboards/rollow/keymaps/via/rules.mk
@@ -1,1 +1,0 @@
-VIA_ENABLE = yes
