+++ keyboards/keyball/keyball39/keymaps/t4corun/readme.md
@@ -3,3 +3,6 @@ cp -r /d/Repo/qmk_userspace/keyboards/keyball /d/Repo/qmk_firmware/keyboards/.
 
 rm -rf /d/Repo/qmk_firmware/keyboards/klor
 cp -r /d/Repo/qmk_userspace/keyboards/klor /d/Repo/qmk_firmware/keyboards/.
+
+rm -rf /d/Repo/qmk_firmware/keyboards/non_tree
+cp -r /d/Repo/qmk_userspace/keyboards/non_tree /d/Repo/qmk_firmware/keyboards/.
