+++ keyboards/non_tree/barbellboards/rollow/keyboard.json
@@ -3,6 +3,9 @@
     "keyboard_name": "Rollow",
     "maintainer": "t4corun",
     "bootloader": "rp2040",
+    "bootmagic": {
+        "enabled": true
+    },
     "diode_direction": "COL2ROW",
     "encoder": {
         "rotary": [
@@ -10,7 +13,6 @@
         ]
     },
     "features": {
-        "bootmagic": true,
         "encoder": true,
         "extrakey": true,
         "oled": true
@@ -46,7 +48,6 @@
         "pid": "0x5257",
         "vid": "0x4242"
     },
-    "community_layouts": ["split_3x5_3", "split_3x6_3"],
     "layouts": {
         "LAYOUT_split_3x5_3": {
             "layout": [
@@ -86,52 +87,6 @@
                 {"matrix": [7, 5], "x": 7, "y": 3.2, "h": 1.5},
                 {"matrix": [7, 4], "x": 8, "y": 3.7},
                 {"matrix": [7, 3], "x": 9, "y": 3.7}
-            ]
-        },
-        "LAYOUT_split_3x6_3": {
-            "layout": [
-                {"matrix": [0, 0], "x": 0, "y": 0.3},
-                {"matrix": [0, 1], "x": 1, "y": 0.3},
-                {"matrix": [0, 2], "x": 2, "y": 0.1},
-                {"matrix": [0, 3], "x": 3, "y": 0},
-                {"matrix": [0, 4], "x": 4, "y": 0.1},
-                {"matrix": [0, 5], "x": 5, "y": 0.2},
-                {"matrix": [4, 5], "x": 9, "y": 0.2},
-                {"matrix": [4, 4], "x": 10, "y": 0.1},
-                {"matrix": [4, 3], "x": 11, "y": 0},
-                {"matrix": [4, 2], "x": 12, "y": 0.1},
-                {"matrix": [4, 1], "x": 13, "y": 0.3},
-                {"matrix": [4, 0], "x": 14, "y": 0.3},
-                {"matrix": [1, 0], "x": 0, "y": 1.3},
-                {"matrix": [1, 1], "x": 1, "y": 1.3},
-                {"matrix": [1, 2], "x": 2, "y": 1.1},
-                {"matrix": [1, 3], "x": 3, "y": 1},
-                {"matrix": [1, 4], "x": 4, "y": 1.1},
-                {"matrix": [1, 5], "x": 5, "y": 1.2},
-                {"matrix": [5, 5], "x": 9, "y": 1.2},
-                {"matrix": [5, 4], "x": 10, "y": 1.1},
-                {"matrix": [5, 3], "x": 11, "y": 1},
-                {"matrix": [5, 2], "x": 12, "y": 1.1},
-                {"matrix": [5, 1], "x": 13, "y": 1.3},
-                {"matrix": [5, 0], "x": 14, "y": 1.3},
-                {"matrix": [2, 0], "x": 0, "y": 2.3},
-                {"matrix": [2, 1], "x": 1, "y": 2.3},
-                {"matrix": [2, 2], "x": 2, "y": 2.1},
-                {"matrix": [2, 3], "x": 3, "y": 2},
-                {"matrix": [2, 4], "x": 4, "y": 2.1},
-                {"matrix": [2, 5], "x": 5, "y": 2.2},
-                {"matrix": [6, 5], "x": 9, "y": 2.2},
-                {"matrix": [6, 4], "x": 10, "y": 2.1},
-                {"matrix": [6, 3], "x": 11, "y": 2},
-                {"matrix": [6, 2], "x": 12, "y": 2.1},
-                {"matrix": [6, 1], "x": 13, "y": 2.3},
-                {"matrix": [6, 0], "x": 14, "y": 2.3},
-                {"matrix": [3, 3], "x": 4, "y": 3.7},
-                {"matrix": [3, 4], "x": 5, "y": 3.7},
-                {"matrix": [3, 5], "x": 6, "y": 3.2, "h": 1.5},
-                {"matrix": [7, 5], "x": 8, "y": 3.2, "h": 1.5},
-                {"matrix": [7, 4], "x": 9, "y": 3.7},
-                {"matrix": [7, 3], "x": 10, "y": 3.7}
             ]
         }
     }
