+++ keyboards/non_tree/barbellboards/rollow/keymaps/t4corun/keymap.c
@@ -1,3 +1,19 @@
+/* Copyright 2024 @t4corun
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #include "t4corun.h"
 
 #define LAYOUT_rollow_wrapper(...) LAYOUT_split_3x5_3(__VA_ARGS__)
