+++ keyboards/non_tree/klor/config.h
@@ -25,7 +25,6 @@
 
 #if defined(OLED_ENABLE)
 #   define OLED_DISPLAY_128X64
-#   define OLED_FONT_H  "./glcdfont.c"
 #endif //OLED_ENABLE
 
 // ┌─────────────────────────────────────────────────┐
