+++ qmk.json
@@ -5,8 +5,8 @@
 		["bastardkb/charybdis/3x5/v2/splinky_3", "t4corun"],
 		["planck/rev6", "t4corun"],
 		["bluebell/swoop", "t4corun"],
-		["barbellboards/rollow", "t4corun"],
-		["klor/saegewerk", "t4corun"],
+		["non_tree/barbellboards/rollow", "t4corun"],
+		["non_tree/klor", "t4corun"],
         ["keyball/keyball39", "t4corun"],
 		["ploopyco/trackball_nano", "t4corun"]
